Move StorageServerDisk implementation to separate files
